requirements.txt
google-adk==1.4.2
streamlit
pandas
openpyxl
easyocr 
Pillow
numpy
xlsxwriter
